Rename paginator file
